Découpage du code sous forme de méthodes et prise de décision en cas de conflit
