+++ CompareSortie
@@ -34,4 +34,3 @@ Z
 7
 8
 9
-
